pyproject.toml
[project]
name = "smtp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.136.1",
    "google-api-python-client>=2.194.0",
    "google-auth>=2.49.2",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "requests>=2.33.1",
    "schedule>=1.2.2",
    "uvicorn>=0.46.0",
]
